backend/requirements.txt
# 核心Flask库
flask
flask-cors          # 处理跨域请求
flask-sqlalchemy    # ORM数据库操作
flask-migrate       # 数据库迁移
Flask-RESTful
Flask-Bcrypt
Flask-Login
Flask-WTF

# 数据库相关
PyMySQL             # MySQL连接库
mysqlclient         # MySQL C扩展，提供更好性能

# 环境变量管理
python-dotenv       # 从.env文件加载环境变量

# 其他常用库
Werkzeug
pytest
coverage
marshmallow
gunicorn
bcrypt              # 密码哈希
# Gunicorn 主要用于 Linux/macOS 系统
gunicorn            # 生产环境WSGI服务器
# Windows
waitress            # 生产环境WSGI服务器
